+++ http/test/test_http.ml
@@ -7,6 +7,12 @@ let handler request =
   match Request.path request with
   | "/error" -> failwith "ERROR"
   | "/echo" -> return (Response.create ~body:(Request.body request) `Ok)
+  | "/no-keep-alive" ->
+    return
+      (Response.create
+         ~headers:(Headers.of_list [ "Connection", "close" ])
+         ~body:(Body.string "This connection will be closed")
+         `Ok)
   | _ -> return (Response.create ~body:(Body.string "Hello World") `Ok)
 ;;
 
@@ -39,12 +45,11 @@ let%expect_test "Simple http endpoint with http client" =
            `POST
            "/hello")
     in
-    printf !"%{sexp: Response.t Or_error.t}" response;
+    printf !"%{sexp: Response.t}" response;
     [%expect
       {|
-    (Ok
-     ((version Http_1_1) (status Ok) (reason_phrase "")
-      (headers ((Content-Length 11))) (body (Fixed "Hello World")))) |}])
+    ((version Http_1_1) (status Ok) (reason_phrase "")
+     (headers ((Content-Length 11))) (body (Fixed "Hello World"))) |}])
 ;;
 
 let%expect_test "Test default error handler" =
@@ -169,11 +174,9 @@ let%expect_test "Client can send streaming bodies" =
                `Repeat (count + 1)))))
     in
     let%bind response =
-      Deferred.Or_error.ok_exn
-        (Client.Oneshot.call
-           (Client.Address.of_host_and_port
-              (Host_and_port.create ~host:"localhost" ~port))
-           (Request.create ~body `POST "/echo"))
+      Client.Oneshot.call
+        (Client.Address.of_host_and_port (Host_and_port.create ~host:"localhost" ~port))
+        (Request.create ~body `POST "/echo")
     in
     let%map body =
       let buf = Buffer.create 32 in
@@ -198,4 +201,231 @@ let%expect_test "Client can send streaming bodies" =
         ((status Ok) (headers ((Transfer-Encoding chunked))) (reason_phrase ""))
 
         Body: "Hello: 1 Hello: 2 Hello: 3 Hello: 4 Hello: 5 " |}])
+;;
+
+let%expect_test "Keep-alives in clients" =
+  Helper.with_server handler ~f:(fun port ->
+    let%bind client =
+      Deferred.Or_error.ok_exn
+        (Client.create
+           (Client.Address.of_host_and_port
+              (Host_and_port.create ~host:"localhost" ~port)))
+    in
+    let body_to_string response =
+      let buf = Buffer.create 32 in
+      let%map () =
+        Body.Stream.iter
+          (Body.to_stream (Response.body response))
+          ~f:(fun v ->
+            Buffer.add_string buf v;
+            Deferred.unit)
+      in
+      Buffer.contents buf
+    in
+    Monitor.protect
+      ~finally:(fun () ->
+        Client.close client;
+        Client.closed client)
+      (fun () ->
+        let%bind response = Client.call client (Request.create `GET "/") in
+        print_s
+          [%sexp
+            { status = (Response.status response : Status.t)
+            ; headers = (Response.headers response : Headers.t)
+            ; reason_phrase = (Response.reason_phrase response : string)
+            }];
+        let%bind body = body_to_string response in
+        printf "\nBody: %S" body;
+        [%expect
+          {|
+    ((status Ok) (headers ((Content-Length 11))) (reason_phrase ""))
+
+    Body: "Hello World" |}];
+        let%bind response =
+          Client.call
+            client
+            (Request.create ~body:(Body.string "This is a body") `POST "/echo")
+        in
+        print_s
+          [%sexp
+            { status = (Response.status response : Status.t)
+            ; headers = (Response.headers response : Headers.t)
+            ; reason_phrase = (Response.reason_phrase response : string)
+            }];
+        let%map body = body_to_string response in
+        printf "\nBody: %S" body;
+        [%expect
+          {|
+    ((status Ok) (headers ((Content-Length 14))) (reason_phrase ""))
+
+    Body: "This is a body" |}]))
+;;
+
+let ensure_aborted fn =
+  Monitor.try_with fn
+  >>= function
+  | Ok response ->
+    failwithf
+      !"Expected request to be aborted, but received a response instead: %{sexp: \
+        Response.t}"
+      response
+      ()
+  | Error exn ->
+    (match Monitor.extract_exn exn with
+     | Client.Request_aborted -> return "Request aborted"
+     | exn -> raise exn)
+;;
+
+let%expect_test "No requests can be sent if a client is closed" =
+  Helper.with_server handler ~f:(fun port ->
+    let%bind client =
+      Deferred.Or_error.ok_exn
+        (Client.create
+           (Client.Address.of_host_and_port
+              (Host_and_port.create ~host:"localhost" ~port)))
+    in
+    let body_to_string response =
+      let buf = Buffer.create 32 in
+      let%map () =
+        Body.Stream.iter
+          (Body.to_stream (Response.body response))
+          ~f:(fun v ->
+            Buffer.add_string buf v;
+            Deferred.unit)
+      in
+      Buffer.contents buf
+    in
+    Monitor.protect
+      ~finally:(fun () ->
+        Client.close client;
+        Client.closed client)
+      (fun () ->
+        let%bind response = Client.call client (Request.create `GET "/") in
+        print_s
+          [%sexp
+            { status = (Response.status response : Status.t)
+            ; headers = (Response.headers response : Headers.t)
+            ; reason_phrase = (Response.reason_phrase response : string)
+            }];
+        let%bind body = body_to_string response in
+        printf "\nBody: %S" body;
+        [%expect
+          {|
+    ((status Ok) (headers ((Content-Length 11))) (reason_phrase ""))
+
+    Body: "Hello World" |}];
+        Client.close client;
+        let%map msg =
+          ensure_aborted (fun () ->
+            Client.call
+              client
+              (Request.create ~body:(Body.string "This is a body") `POST "/echo"))
+        in
+        printf "%s" msg;
+        [%expect {| Request aborted |}]))
+;;
+
+let%expect_test "Clients are automatically closed if Connection:close header is present \
+                 in request"
+  =
+  Helper.with_server handler ~f:(fun port ->
+    let%bind client =
+      Deferred.Or_error.ok_exn
+        (Client.create
+           (Client.Address.of_host_and_port
+              (Host_and_port.create ~host:"localhost" ~port)))
+    in
+    let body_to_string response =
+      let buf = Buffer.create 32 in
+      let%map () =
+        Body.Stream.iter
+          (Body.to_stream (Response.body response))
+          ~f:(fun v ->
+            Buffer.add_string buf v;
+            Deferred.unit)
+      in
+      Buffer.contents buf
+    in
+    Monitor.protect
+      ~finally:(fun () ->
+        Client.close client;
+        Client.closed client)
+      (fun () ->
+        let%bind response =
+          Client.call
+            client
+            (Request.create ~headers:(Headers.of_list [ "Connection", "close" ]) `GET "/")
+        in
+        print_s
+          [%sexp
+            { status = (Response.status response : Status.t)
+            ; headers = (Response.headers response : Headers.t)
+            ; reason_phrase = (Response.reason_phrase response : string)
+            }];
+        let%bind body = body_to_string response in
+        printf "\nBody: %S" body;
+        [%expect
+          {|
+    ((status Ok) (headers ((Content-Length 11))) (reason_phrase ""))
+
+    Body: "Hello World" |}];
+        let%map msg =
+          ensure_aborted (fun () ->
+            Client.call
+              client
+              (Request.create ~body:(Body.string "This is a body") `POST "/echo"))
+        in
+        printf "%s" msg;
+        [%expect {| Request aborted |}]))
+;;
+
+let%expect_test "Clients are automatically closed if Connection:close header is present \
+                 in response"
+  =
+  Helper.with_server handler ~f:(fun port ->
+    let%bind client =
+      Deferred.Or_error.ok_exn
+        (Client.create
+           (Client.Address.of_host_and_port
+              (Host_and_port.create ~host:"localhost" ~port)))
+    in
+    let body_to_string response =
+      let buf = Buffer.create 32 in
+      let%map () =
+        Body.Stream.iter
+          (Body.to_stream (Response.body response))
+          ~f:(fun v ->
+            Buffer.add_string buf v;
+            Deferred.unit)
+      in
+      Buffer.contents buf
+    in
+    Monitor.protect
+      ~finally:(fun () ->
+        Client.close client;
+        Client.closed client)
+      (fun () ->
+        let%bind response = Client.call client (Request.create `GET "/no-keep-alive") in
+        print_s
+          [%sexp
+            { status = (Response.status response : Status.t)
+            ; headers = (Response.headers response : Headers.t)
+            ; reason_phrase = (Response.reason_phrase response : string)
+            }];
+        let%bind body = body_to_string response in
+        printf "\nBody: %S" body;
+        [%expect
+          {|
+    ((status Ok) (headers ((Content-Length 30) (Connection close)))
+     (reason_phrase ""))
+
+    Body: "This connection will be closed" |}];
+        let%map msg =
+          ensure_aborted (fun () ->
+            Client.call
+              client
+              (Request.create ~body:(Body.string "This is a body") `POST "/echo"))
+        in
+        printf "%s" msg;
+        [%expect {| Request aborted |}]))
 ;;